Added logout and JWT token verification
- feature: Implemented POST /api/users/logout route
- auth: Handled user logout
- cookies: Stored JWT token in cookies
- util: Added verifyToken function in src/utils/verifyToken.ts
- verification: Verified JWT token stored in cookies
- security: Secured protected routes by verifying JWT tokens
- doc: Documented logout route and token verification

+++ src/app/api/users/profile/[id]/route.ts
@@ -2,6 +2,7 @@ import { NextResponse, NextRequest } from "next/server";
 import prisma from "@/utils/db";
 import jwt from "jsonwebtoken";
 import { JWTPayload } from "@/utils/types";
+import { verifyToken } from "@/utils/verifyToken";
 
 interface Props {
   params: { id: string };
@@ -22,13 +23,8 @@ export async function DELETE(request: NextRequest, { params }: Props) {
       return NextResponse.json({ message: "user not found" }, { status: 404 });
     }
 
-    const authToken = request.headers.get("authToken") as string;
-    const userFromToken = jwt.verify(
-      authToken,
-      process.env.JWT_SECRET as string
-    ) as JWTPayload;
-
-    if (userFromToken.id === user.id) {
+    const userFromToken = verifyToken(request);
+    if (userFromToken !== null && userFromToken.id === user.id) {
       await prisma.user.delete({ where: { id: parseInt(params.id) } });
 
       return NextResponse.json(
